+++ src/virtual-wan/azext_vwan/aaz/latest/network/p2s_vpn_gateway/__init__.py
@@ -8,3 +8,6 @@
 # pylint: skip-file
 # flake8: noqa
 
+from .__cmd_group import *
+from ._disconnect import *
+from ._reset import *