+++ src/virtual-wan/azext_vwan/aaz/latest/network/vpn_gateway/connection/__cmd_group.py
@@ -8,3 +8,16 @@
 # pylint: skip-file
 # flake8: noqa
 
+from azure.cli.core.aaz import *
+
+
+@register_command_group(
+    "network vpn-gateway connection",
+)
+class __CMDGroup(AAZCommandGroup):
+    """Manage site-to-site VPN gateway connections.
+    """
+    pass
+
+
+__all__ = ["__CMDGroup"]